+++ pages/settings.tsx
@@ -14,6 +14,7 @@ export async function getStaticProps({ locale }: { locale: string }) {
         'common',
         'onboarding',
         'profile',
+        'search',
         'settings',
       ])),
     },
