+++ pages/stats.tsx
@@ -9,6 +9,7 @@ export async function getStaticProps({ locale }: { locale: string }) {
         'common',
         'onboarding',
         'profile',
+        'search',
         'settings',
         'token',
         'trade',
