+++ pages/token/[token].tsx
@@ -9,6 +9,7 @@ export async function getStaticProps({ locale }: { locale: string }) {
       ...(await serverSideTranslations(locale, [
         'common',
         'profile',
+        'search',
         'settings',
         'token',
       ])),
