+++ public/locales/zh_tw/search.json
@@ -1,0 +1,14 @@
+{
+    "mango-account": "Account Address",
+    "mango-account-name": "Account Name",
+    "no-results": "No Results",
+    "no-results-desc": "Try another search. Searches are case-sensitive execpt for profile names which are lowercase",
+    "profile-name": "Profile Name",
+    "results": "Results",
+    "search": "Search",
+    "search-accounts": "Search Accounts",
+    "search-by": "Search By",
+    "search-for": "Search For",
+    "search-failed": "Failed to search",
+    "wallet-pk": "Wallet Address"
+}
